Refactor: hooks
