+++ src/app/[locale]/(pip)/page.client.tsx
@@ -4,12 +4,12 @@ import MoreButton from '@/components/common/button/MoreButton';
 import SearchInput from '@/components/common/input/SearchInput';
 import ChannelSlider from '@/components/home/ChannelSlider';
 import ScheduleSlider from '@/components/home/ScheduleSlider';
-import useCachedData from '@/hooks/useCachedData';
-import useMutateWhitelist from '@/hooks/useDeleteWhitelist';
-import usePostBlacklist from '@/hooks/usePostBlacklist';
-import usePostWhitelist from '@/hooks/usePostWhitelist';
-import useReservePush from '@/hooks/useReservePush';
-import { useScheduleQuery } from '@/hooks/useSchedule';
+import useCachedData from '@/hooks/use-cached-data';
+import useMutateWhitelist from '@/hooks/use-delete-whitelist';
+import usePostBlacklist from '@/hooks/use-post-blacklist';
+import usePostWhitelist from '@/hooks/use-post-whitelist';
+import useReservePush from '@/hooks/use-reserve-push';
+import { useScheduleQuery } from '@/hooks/use-schedule';
 import { Link } from '@/libraries/i18n';
 import { useLocale, useTranslations } from '@/libraries/i18n/client';
 import { TParsedClientContent } from '@/libraries/mongodb/type';
