+++ src/app/[locale]/(pip)/page.tsx
@@ -1,6 +1,6 @@
 import { serverApi } from '@/apis/fetcher';
 import Footer from '@/components/common/Footer';
-import { CHANNELS_TAG, METADATAS_TAG } from '@/constants/revalidateTag';
+import { CHANNELS_TAG, METADATAS_TAG } from '@/constants/revalidate-tag';
 import { TYChannelReturn } from '@/libraries/mongodb/channels';
 import { TMetadata } from '@/types';
 import Client from './page.client';
