+++ src/components/schedule/Home.tsx
@@ -1,6 +1,6 @@
 'use client';
-import useCachedData from '@/hooks/useCachedData';
-import { useScheduleQuery } from '@/hooks/useSchedule';
+import useCachedData from '@/hooks/use-cached-data';
+import { useScheduleQuery } from '@/hooks/use-schedule';
 import { useLocale, useTranslations } from '@/libraries/i18n/client';
 import { TScheduleDto } from '@/types/dto';
 import { addEscapeCharacter } from '@/utils/regexp';
