+++ src/components/setlist/Table.tsx
@@ -4,8 +4,8 @@ import Nodata from '@/components/common/Nodata';
 import loadingCss from '@/components/common/loading/Loading.module.scss';
 import Wave from '@/components/common/loading/Wave';
 import { SETLIST_PAGE_SIZE } from '@/constants';
-import { SETLISTS_TAG } from '@/constants/revalidateTag';
-import useCachedData from '@/hooks/useCachedData';
+import { SETLISTS_TAG } from '@/constants/revalidate-tag';
+import useCachedData from '@/hooks/use-cached-data';
 import { useLocale, useTranslations } from '@/libraries/i18n/client';
 import type { Setlist } from '@/libraries/oracledb/setlist/service';
 import type { GetSetlistRes } from '@/types/api/setlist';
