+++ src/hooks/use-delete-whitelist.ts
@@ -1,16 +1,15 @@
 import { clientApi } from '@/apis/fetcher';
-import { BLACKLIST_TAG } from '@/constants/revalidateTag';
+import { WHITELIST_TAG } from '@/constants/revalidate-tag';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { Session } from 'next-auth';
 import { toast } from 'sonner';
 
-export default function usePostBlacklist() {
+export default function useMutateWhitelist() {
   const queryClient = useQueryClient();
-
-  const mutation = useMutation({
+  return useMutation({
     mutationFn: async ({ session, channelId }: { session: Session; channelId: string }) => {
       const json = await clientApi
-        .post<{ message: string; data: string }>(`v1/blacklist/${channelId}`, {
+        .delete<{ message: string; data: string }>(`v1/whitelist/${channelId}`, {
           headers: { Authorization: `Bearer ${session.user.accessToken}` },
         })
         .json();
@@ -18,17 +17,15 @@ export default function usePostBlacklist() {
     },
     onSuccess: (res) => {
       toast.success(res.message);
-      if (queryClient.getQueryData([BLACKLIST_TAG])) {
-        queryClient.setQueryData([BLACKLIST_TAG], (prev: string[]) => {
-          return [...prev, res.data];
+      if (queryClient.getQueryData([WHITELIST_TAG])) {
+        queryClient.setQueryData([WHITELIST_TAG], (prev: string[]) => {
+          return prev.filter((id) => id !== res.data);
         });
       }
     },
     onError: () => {
       toast.error('서버에러가 발생했습니다. 잠시후 다시 시도해주세요.');
-      queryClient.invalidateQueries({ queryKey: [BLACKLIST_TAG] });
+      queryClient.invalidateQueries({ queryKey: [WHITELIST_TAG] });
     },
   });
-
-  return mutation;
 }
