+++ src/hooks/use-post-blacklist.ts
@@ -1,32 +1,32 @@
 import { clientApi } from '@/apis/fetcher';
-import { WHITELIST_TAG } from '@/constants/revalidateTag';
+import { BLACKLIST_TAG } from '@/constants/revalidate-tag';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { Session } from 'next-auth';
 import { toast } from 'sonner';
 
-export default function usePostWhitelist() {
+export default function usePostBlacklist() {
   const queryClient = useQueryClient();
 
   const mutation = useMutation({
     mutationFn: async ({ session, channelId }: { session: Session; channelId: string }) => {
-      const response = await clientApi
-        .post<{ message: string; data: string }>(`v1/whitelist/${channelId}`, {
+      const json = await clientApi
+        .post<{ message: string; data: string }>(`v1/blacklist/${channelId}`, {
           headers: { Authorization: `Bearer ${session.user.accessToken}` },
         })
         .json();
-      return response;
+      return json;
     },
     onSuccess: (res) => {
       toast.success(res.message);
-      if (queryClient.getQueryData([WHITELIST_TAG])) {
-        queryClient.setQueryData([WHITELIST_TAG], (prev: string[]) => {
+      if (queryClient.getQueryData([BLACKLIST_TAG])) {
+        queryClient.setQueryData([BLACKLIST_TAG], (prev: string[]) => {
           return [...prev, res.data];
         });
       }
     },
     onError: () => {
       toast.error('서버에러가 발생했습니다. 잠시후 다시 시도해주세요.');
-      queryClient.invalidateQueries({ queryKey: [WHITELIST_TAG] });
+      queryClient.invalidateQueries({ queryKey: [BLACKLIST_TAG] });
     },
   });
